fixed some edge cases for ft_strnstr and changed int to size_t variables where applicable

+++ libft/ft_strchr.c
@@ -14,9 +14,9 @@
 
 char	*ft_strchr(const char *s, int c)
 {
-	int		s_l;
-	int		len;
-	char	*tmp;
+	size_t		s_l;
+	size_t		len;
+	char		*tmp;
 
 	tmp = (char*)s;
 	s_l = 0;
